refactor: reports/ui 컴포넌트를 common/ 디렉토리로 이동
