+++ pages/admin/reports/reviews.js
@@ -4,7 +4,7 @@ import ReviewReportTable from "@/components/reports/tables/ReviewReportTable";
 import ReviewReportDetail from "@/components/reports/details/ReviewReportDetail";
 import { onAuthStateChanged } from "firebase/auth";
 import { auth } from "@/lib/firebase/client";
-import NoReportSelected from "@/components/reports/ui/NoReportSelected";
+import NoReportSelected from "@/components/common/NoReportSelected";
 
 export default function AdminReviewReportsPage() {
   const [authReady, setAuthReady] = useState(false);
